added utils.c removed scalar operations from ops
