+++ mithril/cores/c/raw_c/compile.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-cc ops.c array.c -shared -fPIC -o libmithrilc.so
+cc ops.c array.c utils.c -shared -fPIC -g -o libmithrilc.so
